configure root store & reducer, convert mainReducer into a slice
